Migrated routes to alts.ts
